requirements.txt
fastapi==0.73.0
uvicorn==0.17.1
gspread==5.1.1
oauth2client==4.1.3
aiohttp==3.8.1
Pillow==9.0.0
pydantic==1.9.0
